TEC-4761 Add checkstyle and apache 2 licence header
