--- src/main/java/com/echobox/api/linkedin/types/CodeAndNameType.java
@@ -1,21 +1,0 @@
-package com.echobox.api.linkedin.types;
-
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * Code and name model wrapper
- * @author Joanna
- *
- */
-public class CodeAndNameType {
-  
-  @Getter
-  @Setter
-  private String code;
-  
-  @Getter
-  @Setter
-  private String name;
-
-}
