--- src/main/java/com/echobox/api/linkedin/types/ContactInfo.java
@@ -1,33 +1,0 @@
-package com.echobox.api.linkedin.types;
-
-import lombok.Getter;
-import lombok.Setter;
-
-/**
- * Contact info model
- * @author Joanna
- *
- */
-public class ContactInfo {
-  
-  /**
-   * Company phone number for the location.
-   */
-  @Getter
-  @Setter
-  private String phoneOne;
-
-  /**
-   * Second company phone number for the location.
-   */
-  @Getter
-  @Setter
-  private String phoneTwo;
-
-  /**
-   * Company fax number for the location.
-   */
-  @Getter
-  @Setter
-  private String fax;
-}
